Update README to reflect provider release
Now that the provider has been released, there are a number of items in
the old README that are no longer relevant. This commit replaces the
README with one that has been template from the vSphere provider, and
should better reflect what one would need to do to use the provider or
develop it, including up to date installation and development
instructions, and a referral to the Terraform website for documentation.