Mark edge_cluster_id as ForceNew in T1 router
Backend does not allow changing edge cluster with single update call,
we would need to disable firewall first, then dissassociate the cluster
and associate the new one. Until this is implemented, use ForceNew as
workaround.
This seems to trigger a mismatch bug on terraform apply, but the bug
goes away if we apply again. Will open the bug upstream.