Pass IDs for the FW sections and rules for update
Currently when updating a firewall rules internally they get
deleted and the new ones are created (causing their IDs to change).
This is not always a desired behaviour, as the NSX-T API itself
allows changing the rules in place.

This patch adds the current firewall section and rule ID to the
payload in order to allow preserving the IDs.