fix refresh failures
The nsxt resources do not behave correctly during the refresh cycle,
if the described resource has vanished. Here an error 404 is reported
and terraform stops processing (instead of recreating the missing resource).

This is caused by a general implementation error in nearly all
resources based on a wrong assumption about the behaviour of the go SDK.
It assumes that http status codes are never reported as errors, but
all the SDK functions return an error if the status code is above 300.
Therefore the order of the error handling in the read functions of
the resources is wrong.

Instead of

```
if err != nil {
  return fmt.Errorf("Error during LogicalSwitch read: %v", err)
}
if resp.StatusCode == http.StatusNotFound {
  log.Printf("[DEBUG] LogicalSwitch %s not found", id)
  d.SetId("")
  return nil
}
```

the correct error handling must always be

```
if resp != nil && resp.StatusCode == http.StatusNotFound {
  log.Printf("[DEBUG] LogicalSwitch %s not found", id)
  d.SetId("")
  return nil
}
if err != nil {
  return fmt.Errorf("Error during LogicalSwitch read: %v", err)
}
```

In three cases the order was already fixed, but in two of these cases
the nil pointer check was missing, causing a NPE for http request
execution errors.