Add imported for predefined gateway policy
VMC auto-creates a rule under default policy. In order to provide best
experience for the user, we give an option to import this rule into
terraform state.
However, the user is not required to import the policy, if the
predefined rule is present in terraform config. This change also
expands the doc to cover both approaches.