Determine NSX version in VMC environment
Some VMC customers might remain with NSX 2.5 for a while. Due to
significant changes in support for NSX below and above 3.0.0, we
need a way to distinguish between the two. Since version API is
not accessible in VMC, we use search API which is not available
before 3.0.0.
In future, a more reliable way is to be introduces.