Update FW policy rules only if needed
In security policy and gateway policy resources, any change
would trigger an update for the whole list of rules. This change
switches to use H-API, and would update rules on individual basis,
and rule will be updated only if terraform detects a change in it.
This will prevent needless reset of statistics for untouched rules.

Please note that if a rule is deleted, terraform would detect a
change in all subsequent rules, since rules are considered an
ordered list. User can consider disabling rule as an alternative.

Signed-off-by: Anna Khmelnitsky <akhmelnitsky@vmware.com>