Fix mixup in condition to address lb rules
A negation was missing in assesment of rule change, that relates to
provider version upgrade workaround. When no change in rules is
detected, we assume current rules should be preserved, for sake of
rules that were added manually on NSX when rule feature was not
present in proveder.

In addition, fix inconsistent documentation.
Signed-off-by: Anna Khmelnitsky <akhmelnitsky@vmware.com>