Support on-demand connection init in the provider
When the new `on_demand_connection` provider flag is set to true,
provider will not verify connection and pull NSX version on provider
startup. Instead, those operations will be done for each provider
instance when terraform objects are evaluated.
This setting is useful when NSX appliance is being deployed via same
apply process, and not yet responsive when the provider is initialized.
This setting is not supported for VMC and not compatible with deprecated
objects implemented via the old SDK. It is also temporarily incompatible
with license feature (this will be fixed in follow up PR)

Signed-off-by: Anna Khmelnitsky <akhmelnitsky@vmware.com>