Fix rule order bug
NSX can not guarantee rule order unless sequence_number is provided, or
`revise` API is called.
This change auto-assigns sequence number to all rules in list, for which
it was not provided by the user.
It takes care of provider upgrade by making sure rules without assigned
sequence number are updated even if terraform does not detect a change
in them.

Signed-off-by: Anna Khmelnitsky <akhmelnitsky@vmware.com>