Avoid applying lisence on every init
Previous code assumed GetChange SDK routine works for provider
attributes, which it doesn't (provider attributes are not stored in
state).
In order to avoid reapplying licenses on every init, this change
pulls license keys from backend and only applies the diff.
The provider will never remove licenses, this is to avoid scenario
when provider removes license that was manually applied on NSX.

Signed-off-by: Anna Khmelnitsky <akhmelnitsky@vmware.com>