Add Principle Identity resource (#1021)
This PR adds support of Principle Identity resource. Due to limitations
in API, PI in terraform is treated as immutable resource. Any changes
to the definition will force a resource recreation.

This change also refactos role schema. Now for both PI and role
bindings, roles for a path can be defined as a list, rather than sub
resources for convenience.

Signed-off-by: Shawn Wang <wshaoquan@vmware.com>