Change password for BGP neighbor only on demand
Since password is not sent back from NSX for security reasons,
it will not be present in state after import.
In this scenario, we need to avoid the password from being cleared
next time an update is triggered on the imported resource.
Hence we only update the password if a change in this field s detected.
Note that user can leave the password blank in configuration of the
imported resource.

Signed-off-by: Anna Khmelnitsky <akhmelnitsky@vmware.com>