Handle password change in update
NSX requires sending the old password while changing password.
So the change uses the old password from the state, to verify that
password change is legal.

There is an API to reset password - it requires an admin
permission and won't check old password. IMO it's better so as otherwise
it would indicate that state is not in sync with NSX.

Signed-off-by: Kobi Samoray <kobi.samoray@broadcom.com>