Avoid returning nil object in StateRefreshFunc
When defining state refresh functions for terraform sdk wait
mechanism, we normally return nil for object while waiting for
state. This is actually 'not found' indication for sdk code, and
sdk implements an additional stop mechanism for such cases, which
causes the wait loop to exit prematurely before timeout is
reached. For example, in host resouce, this would cause 'failed to
remove nsx' error on delete.
In order to avoid for this mechanism to kick in, the code is changed
to return 'notyet' string rather then nil object in the wait func.

Signed-off-by: Anna Khmelnitsky <akhmelnitsky@vmware.com>