Fix edge transport node state func
When defining state refresh functions for terraform sdk wait
mechanism, we normally return nil for object while waiting for state.
This is actually 'not found' indication for sdk code, and sdk
implements an additional stop mechanism for such cases, which causes
the wait loop to exit  prematurely before timeout is reached.
This fix was made for all resources in previous commit.
This change completes the fix for edge transport node, which had its
delete wait mechanism introduced recently and thus does not contain
the fix.

Signed-off-by: Anna Khmelnitsky <akhmelnitsky@vmware.com>