order realized VM results by external_id
This PR adds external_id as the sortByParam for calls to the realized VM
endpoint. This is needed in larger environments (1k+ VMs) because as of
NSX-T 3.2.3 the /policy/api/v1/infra/realized-state/virtual-machines
endpoint is not respecting a consistent ordering by default. This means
that the internal slice representing VMs returned from
listAllPolicyVirtualMachines may include duplicates and may be missing
VMs that actually exist. Explicitly setting an order forces consistent
results across the paginated calls.

Resolves: #1193
Signed-off-by: Paul Zerkel <zerkel@gmail.com>