Wait for ip address allocation on create
When ip address allocation resource is used without specific IP,
platform allocates the address from the pool asyncronously.
Previously, wait for the allocation happened in Read flow. This is
problematic, since if allocation fails, resource cannot be deleted
unless -refresh=false option is used.
This change fixes the code to wait for allocation in create flow.
If allocation fails, create fails as well.

Signed-off-by: Anna Khmelnitsky <anna.khmelnitsky@broadcom.com>