fix: staticcheck QF1001
Modifies the conditional expression in `data_source_nsxt_policy_transport_zone.go`
to address the QF1001 staticcheck warning.

The diff of the expressions are logically equivalent according to De Morgan's law, which states that !(A && B) is equivalent to (!A || !B). This change maintains identical behavior while
expressing the condition in a more direct and easier-to-understand form that aligns with staticcheck.

Signed-off-by: Ryan Johnson <ryan.johnson@broadcom.com>