Handle non-triggered upgrade prechecks
A user specifies a list of prechecks to be acknowledged.
Each of these prechecks is attempted to be acknowledged, but the acknowledge resource
does not verify that the precheck actually failed.
This change will verify that the precheck actually failed before acknowledging it.

Also verify that all precheck ids provided are valid.

Signed-off-by: Kobi Samoray <kobi.samoray@broadcom.com>